refactor(metrics): wire consumers to the SPI's new module
Import-only changes to point every remaining consumer at
org.apache.polaris.spi.substrate.metrics: NoOpMetricsReporter,
PersistingMetricsReporter, IcebergCatalogHandler,
IcebergCatalogHandlerFactory, ServiceProducers, and
IcebergCatalogHandlerTest. None of these change behavior; selection
logic in ServiceProducers.metricsReporter(...) is untouched.

extensions/metrics-reports/persistence/relational-jdbc no longer needs
runtime/service either, for the same reason as the base module. Both
extensions and runtime/service now take a dependency on the spi module
for the contract. runtime/service needs it on testFixtures too, because
TestServices supplies a lambda reporter to the handler builder, so the
interface has to resolve there as well.

runtime/service previously carried its default reporter
(LoggingMetricsReporter) directly, so it never needed to depend on the
metrics extension modules. Now that the default lives in
extensions/metrics-reports/base, add it back as runtimeOnly: the
packaged Quarkus app needs it on its runtime classpath for bean
discovery, same as the existing runtimeOnly(":polaris-relational-jdbc")
line right above it. runtime/server already carries this same
dependency for the assembled distribution.
