+++ extensions/metrics-reports/persistence/relational-jdbc/build.gradle.kts
@@ -24,8 +24,8 @@ plugins {
 
 dependencies {
   implementation(project(":polaris-core"))
+  implementation(project(":polaris-spi"))
   implementation(project(":polaris-relational-jdbc"))
-  implementation(project(":polaris-runtime-service"))
   implementation(project(":polaris-extensions-metrics-reports"))
 
   implementation(platform(libs.iceberg.bom))
