+++ runtime/service/src/main/java/org/apache/polaris/service/catalog/iceberg/IcebergCatalogHandler.java
@@ -132,10 +132,10 @@ import org.apache.polaris.service.http.IcebergHttpUtil;
 import org.apache.polaris.service.http.IfNoneMatch;
 import org.apache.polaris.service.idempotency.EntityIdempotency;
 import org.apache.polaris.service.idempotency.IdempotencyRequestContext;
-import org.apache.polaris.service.metrics.IcebergMetricsReporter;
-import org.apache.polaris.service.metrics.MetricType;
-import org.apache.polaris.service.metrics.MetricsReportEnvelope;
 import org.apache.polaris.service.types.NotificationRequest;
+import org.apache.polaris.spi.substrate.metrics.IcebergMetricsReporter;
+import org.apache.polaris.spi.substrate.metrics.MetricType;
+import org.apache.polaris.spi.substrate.metrics.MetricsReportEnvelope;
 import org.jspecify.annotations.NonNull;
 import org.jspecify.annotations.Nullable;
 import org.slf4j.Logger;
