+++ runtime/service/src/main/java/org/apache/polaris/service/catalog/iceberg/IcebergCatalogHandlerFactory.java
@@ -39,7 +39,7 @@ import org.apache.polaris.service.catalog.io.StorageAccessConfigProvider;
 import org.apache.polaris.service.config.ReservedProperties;
 import org.apache.polaris.service.events.EventAttributeMap;
 import org.apache.polaris.service.idempotency.IdempotencyRequestContext;
-import org.apache.polaris.service.metrics.IcebergMetricsReporter;
+import org.apache.polaris.spi.substrate.metrics.IcebergMetricsReporter;
 
 @RequestScoped
 public class IcebergCatalogHandlerFactory {
